requirements.txt
Flask==3.0.2
openai==1.16.1
python-dotenv==1.0.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pyOpenSSL==24.2.1
